requirements.txt
qrcode==7.4.2
quart==0.18.4
python-telegram-bot==20.4
urllib3==1.23
